% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_mer_data}
\alias{get_mer_data}
\title{MER}
\usage{
get_mer_data(api_key, compnum)
}
\arguments{
\item{api_key}{API Key object (see also, setapi())}

\item{compnum}{company number (see also, compdata)}
}
\description{
MER
}
