% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_mer_data}
\alias{get_mer_data}
\title{MER}
\usage{
get_mer_data(api_key, compnum)
}
\arguments{
\item{api_key}{API Key object (see also, setapi())}

\item{compnum}{company number (see also, compdata)}
}
\description{
https://www.quandl.com/databases/MF1/documentation
Mergent Global Fundamental Data에서 제공한 데이터를 검색할 수 있습니다.
Quandly의 api_key를 첫번째 인자로, compdata의 compnumber를 두번째 인자로 입력합니다.
}
